Arithmetic update
Renamed ArithmeticScope to ArithmeticQueue, since Queue is more accurate
to what it is doing that ArithmeticScope.

Added ArithmeticNumeric struct which abstracts all arithmetic functions
for all of the integer/decimal types allowing for different numerical
types being operanded with other numerical types. (e.g byte addition on
double).
This struct will solve the issue in the ArithmeticQueue, which we had to
use the double type on everything which would use alot of memory for
something that could just be 8bits wide.