Arithmetic functions
Now added support for functions in the ArithmeticQueue.
ArithmeticFunction allows you to create your own function via delegate
OR override the entire class for faster performance.

When the ArithmeticFunction's Call() is called from the ArithmeticQueue
Calculate function, the method is invoked which is passed the result of
the queue which was parsed as the functions arguments.
The function then returns a ArithmeticNumeric struct which during
calculation gets substituted into the calculation.