CSS Pseudo Class - Nth*
PseudoClass_Nth class constructor now processes an expression and
deturmines how we make n the subject by using known CSS nth patterns (n,
an, an+b).

We need to make n the subject for optimization on when we find if an
element should be selected or not, to see if the index matches one of
the results of the expression. If we already knew what N is, we can know
which index's will be matched.

(If N does not match, the output of the reversed expression will always
be decimal and not integral)